[refactor/#63] 그외 컴포너트 구조 개선
