+++ src/components/my-activities/registration/ExperienceForm.tsx
@@ -6,13 +6,12 @@ import { ChangeEvent, useRef, useState } from "react";
 import { TimeSlotsEditor } from "./TimeSlotsEditor";
 import { BannerImagePicker } from "./BannerImagePicker";
 import { IntroImagesPicker } from "./IntroImagesPicker";
-import { MenuItem } from "@/components/ui/Dropdown";
+import { MenuItem } from "@/components/ui/Dropdown/Dropdown";
 import { TextInput } from "@/components/ui/Input/TextInput";
 import { SelectInput } from "@/components/ui/Input/SelectInput";
 
 // Icons
 import example from "@/assets/svgs/example.svg";
-
 
 const items: MenuItem[] = [
   { label: "액티비티" },
