+++ src/components/ui/Modal/ExperienceReservationModal.tsx
@@ -1,11 +1,15 @@
 "use client";
 
-import { NumberStepper } from "./NumberStepper";
-import CloseIcon from "@/assets/svgs/close_icon.svg";
 import Image from "next/image";
 import { Dialog } from "@headlessui/react";
 import { ModalProps } from "@/types/modalProps";
+
+// UI
 import { SchedulePicker } from "@/components/activity/SchedulePicker";
+
+// Icons
+import CloseIcon from "@/assets/svgs/close_icon.svg";
+import { NumberStepper } from "@/components/activity/NumberStepper";
 
 const slots = [
   { date: "2025-09-01", startTime: "12:00", endTime: "13:00" },
